:recycle: move shutdown handling to launcher
