+++ vllm/entrypoints/api_server.py
@@ -116,6 +116,7 @@ async def run_server(args: Namespace,
 
     shutdown_task = await serve_http(
         app,
+        engine=engine,
         host=args.host,
         port=args.port,
         log_level=args.log_level,
